refactor: 메신저 화면, 채팅방 화면 별도의 창으로 분리
